Add generic show dialog to centralize tabbase and wizard login
