+++ extensions/sql-migration/src/dashboard/tabBase.ts
@@ -12,6 +12,7 @@ import { DatabaseMigration } from '../api/azure';
 import { getSelectedServiceStatus } from '../models/migrationLocalStorage';
 import { MenuCommands, SqlMigrationExtensionId } from '../api/utils';
 import { DashboardStatusBar } from './DashboardStatusBar';
+import { ShowStatusMessageDialog } from '../dialog/generic/genericDialogs';
 
 export const EmptySettingValue = '-';
 
@@ -202,49 +203,6 @@ export abstract class TabBase<T> implements azdata.Tab, vscode.Disposable {
 		statusMessage: string,
 		errorMessage: string,
 	): void {
-		const tab = azdata.window.createTab(title);
-		tab.registerContent(async (view) => {
-			const flex = view.modelBuilder.flexContainer()
-				.withItems([
-					view.modelBuilder.text()
-						.withProps({ value: statusMessage })
-						.component(),
-				])
-				.withLayout({
-					flexFlow: 'column',
-					width: 420,
-				})
-				.withProps({ CSSStyles: { 'margin': '0 15px' } })
-				.component();
-
-			if (errorMessage.length > 0) {
-				flex.addItem(
-					view.modelBuilder.inputBox()
-						.withProps({
-							value: errorMessage,
-							readOnly: true,
-							multiline: true,
-							inputType: 'text',
-							height: 100,
-							CSSStyles: { 'overflow': 'hidden auto' },
-						})
-						.component()
-				);
-			}
-
-			await view.initializeModel(flex);
-		});
-
-		const dialog = azdata.window.createModelViewDialog(
-			title,
-			'messageDialog',
-			450,
-			'normal');
-		dialog.content = [tab];
-		dialog.okButton.hidden = true;
-		dialog.cancelButton.focused = true;
-		dialog.cancelButton.label = loc.CLOSE;
-
-		azdata.window.openDialog(dialog);
+		ShowStatusMessageDialog(title, statusMessage, errorMessage);
 	}
 }
