+++ extensions/sql-migration/src/models/migrationWizardPage.ts
@@ -5,7 +5,8 @@
 
 import * as azdata from 'azdata';
 import { MigrationStateModel, StateChangeEvent } from './stateMachine';
-import * as constants from '../constants/strings';
+import { ShowStatusMessageDialog } from '../dialog/generic/genericDialogs';
+
 export abstract class MigrationWizardPage {
 	constructor(
 		protected readonly wizard: azdata.window.Wizard,
@@ -87,49 +88,6 @@ export abstract class MigrationWizardPage {
 		statusMessage: string,
 		errorMessage: string,
 	): void {
-		const tab = azdata.window.createTab(title);
-		tab.registerContent(async (view) => {
-			const flex = view.modelBuilder.flexContainer()
-				.withItems([
-					view.modelBuilder.text()
-						.withProps({ value: statusMessage })
-						.component(),
-				])
-				.withLayout({
-					flexFlow: 'column',
-					width: 420,
-				})
-				.withProps({ CSSStyles: { 'margin': '0 15px' } })
-				.component();
-
-			if (errorMessage.length > 0) {
-				flex.addItem(
-					view.modelBuilder.inputBox()
-						.withProps({
-							value: errorMessage,
-							readOnly: true,
-							multiline: true,
-							inputType: 'text',
-							height: 100,
-							CSSStyles: { 'overflow': 'hidden auto' },
-						})
-						.component()
-				);
-			}
-
-			await view.initializeModel(flex);
-		});
-
-		const dialog = azdata.window.createModelViewDialog(
-			title,
-			'messageDialog',
-			450,
-			'normal');
-		dialog.content = [tab];
-		dialog.okButton.hidden = true;
-		dialog.cancelButton.focused = true;
-		dialog.cancelButton.label = constants.CLOSE;
-
-		azdata.window.openDialog(dialog);
+		ShowStatusMessageDialog(title, statusMessage, errorMessage);
 	}
 }
